+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - handling of brackets in the changelog version numbers
 - Reissue::Parser class to parse the changelog file
 - Reissue::Printer class to print the changelog file
+- Reissue::Task class to handle the creation of reissue tasks
 
 ### Fixed: